+++ metplus/wrappers/pb2nc_wrapper.py
@@ -224,62 +224,3 @@ class PB2NCWrapper(LoopTimesWrapper):
             cmd += f" -valid_end {self.c_dict['VALID_WINDOW_END']}"
 
         return cmd
-
-    def get_files_from_time(self, time_info):
-        """! Create dictionary containing time information (key time_info) and
-             any relevant files for that runtime. The parent implementation of
-             this function creates a dictionary and adds the time_info to it.
-             This wrapper gets all files for the current runtime and adds it to
-             the dictionary with keys 'FCST' and 'OBS'
-
-             @param time_info dictionary containing time information
-             @returns dictionary containing time_info dict and any relevant
-             files with a key representing a description of that file
-        """
-        var_list = [None]
-
-        # create a dictionary for each field (var) with time_info and files
-        file_dict_list = []
-        for var_info in var_list:
-            if var_info:
-                add_field_info_to_time_info(time_info, var_info)
-            file_dict = super().get_files_from_time(time_info)
-            file_dict['var_info'] = var_info
-
-            input_files, offset_time_info = (
-                self.get_input_files(time_info, fill_missing=True)
-            )
-            file_dict['time_info'] = offset_time_info
-
-            # only add all input files if none are missing
-            no_missing = True
-            if input_files:
-                for key, value in input_files.items():
-                    if 'missing' in value:
-                        no_missing = False
-                    file_dict[key] = value
-            if no_missing:
-                file_dict_list.append(file_dict)
-
-        return file_dict_list
-
-    def _update_list_with_new_files(self, time_info, list_to_update):
-        new_files = self.get_files_from_time(time_info)
-        if not new_files:
-            return
-
-        # if list to update is empty, copy new items into list
-        if not list_to_update:
-            for new_file in new_files:
-                list_to_update.append(new_file.copy())
-            return
-
-        # if list to update is not empty, add new files to each file list,
-        # make sure new files correspond to the correct field (var)
-        assert len(list_to_update) == len(new_files)
-        for new_file, existing_item in zip(new_files, list_to_update):
-            assert new_file['var_info'] == existing_item['var_info']
-            for key, value in new_file.items():
-                if key == 'var_info' or key == 'time_info':
-                    continue
-                existing_item[key].extend(value)
